feat: Permissive deserializer
