+++ Cargo.toml
@@ -32,3 +32,4 @@ serde_yaml = "0.9.34"
 thiserror = "1.0.61"
 clap = { version = "4.5", features = ["derive"], optional = true }
 schemars = { version = "0.8.21", optional = true }
+regex = "1.10.5"
